Restructured text filter handler.
